Fix DeserializingEventJournalMapEvent.toString
This class did not and still does not implement toString(), it inherits
it form `InternalEventJournalMapEvent`, which contained its class name
in the string. This was misleading as normally the user deals with the
`DeserializingEventJournalMapEvent` subclass. When the user prints it
(such as when developing a Jet job), it can be misleading.