Fix snapshot creation after execution is done (#1397)
For non-distributed sources that use `forceTotalParallelismOne`, batch
processors are created on inactive members. If there's no distributed
edge in the DAG, the execution on those members will complete
immediately, while the execution on the active members tries to do
snapshot.

This PR changes the edge case behavior that if the execution is
completed, we immediately respond to the SnapshotOperation instead of
throwing a CancellationException.