Close resources in config builders (#15027)
The recently written config resolution tests failed on Windows because
some config files created during the test remained open and due to this
these files on Windows couldn't be deleted at the end of the test,
failing the subsequent tests. This commit fixes the unclosed resources
issues in two places.

1. Makes AbstractYamlConfigBuilder and YamlConfigBuilder to close the 
resources it loaded the YAML document from.
2. Fixed XMLClientFailoverBuilder to call locateEverywhere() method of
the used locator only if the provided locator instance is null and a
newly created locator is being used. Otherwise the locator logic is
executed twice and the config file gets open twice, but only one of the
resources are closed.

Fixes #14988