Improve performance of TransactionLog.add() by avoiding LinkedList.remove() call.
Code may be considered cleaner/simpler now, but the reason for this change is performance. When in a single transaction entries were changed, and then changed again (TransactionalMap.put() called on the same entry), the TransactionLog.add() method did a LinkedList.remove() call which has complexity O(n). The new implementation that uses LinkedHashMap instead of LinkedList+HashMap is up to 20% faster on 1000-entry transaction, and up to three times faster on 10000-entry transaction. On small transactions performance is the same.