IMap entries eviction support on the indexes level (#11748)
For on-heap indexes update record's lastAccessTime when it is being accessed
through the index, so that expiration maxIdle logic takes this into account.

An index still may return stale entries, because expiration logic is not thread
safe and runs on the partition thread. However, the impact is limited because
a periodic cleaning task evicts entries anyway.

Part of (#11748)