Fix loss of SnapshotOperation (#1435)
The previous fix in #1377 wasn't sufficient, it only worked when there
was no distributed edge in the DAG. A processor after a distributed edge
got stuck forever waiting for the barrier from a member that did not
receive the SnapshotOp.

The fix here is to retry the SnapshotOperation. If the operation was
lost, the snapshot will eventually succeed. If the response was lost,
even though the snapshot operation is not idempotent and will fail when
retried with the same snapshotId, at least the job won't get stuck.
Snapshot failure is then ignored.

Note that the test does not test when a response to an operation is lost,
there is currently no test harness to simulate that.