Move MerkleTreeConfig inside MapConfig
Initially, merkle tree configuration was put as a top-level config
because we were anticipating its' usage in cache as well but we were not
able to change CacheConfig because of client backwards compatibility
(see BinaryInterface annotation).
Now with 4.0, we will remove this restriction which means merkle tree
config can be part of both map and cache config (once it's implemented
for cache).
Using it this way will eliminate some of the confusion as up until now
the user had to additionally specify the mapName on the merkle tree
configuration to 'connect' the merkle tree configuration to the map
configuration.