Move EventJournalConfig inside MapConfig and CacheConfig
Initially, event journal configuration was put as a top-level config
because we are not able to change CacheConfig because of client
backwards compatibility (see BinaryInterface annotation).
Now with 4.0, we will remove this restriction which means event journal
config can be part of both map and cache config.
Using it this way will eliminate some of the confusion as up until now
the user had to additionally specify the mapName and cacheName on the
event journal configuration to 'connect' the event journal configuration
to the map configuration.