Rewrite JavaVersion (#15372)
Java9 introduced java.lang.Runtime.Version with a major() method that
returns the major version as an int that the replaced code tried to
figure out by parsing the java.version system property. Since with 4.0
Hazelcast requires at least Java8 runtime, it's easy to detect if the
runtime is Java8: this class is not accessible via reflection.
Otherwise we can read out the major version via the mentioned method.
Since this is part of the official API dealing with unknown Java
versions is not an issue anymore, we can safely compare expected
versions against the runtime version. For this purpose
JavaVersion.FutureJavaVersion inner class and the JavaMajorVersion
interface is added.

In scope of this work ClassloadingMutexProvider got removed. This
provider checked if parallel class loading was in use (if Java7+ version)
and returned a mutex safe to be used with the class loading mechanism.
With Hazelcast 4.0 and Java8 we always take the Java7+ branch, therefore
no need for this class.

Implements #15348