com.hz.spi classes/interface privateapi
All the classes from the com.hz.spi package are made privateapi.

So if we are not ready in time cleaning up the com.hz.spi, these
classes are not for external consumption and therefor we can safely
move them in a later release.

We do not expect end users to write custom services. This was a goal
in 3.0, but not any longer.