Implement the parallel leader & follower Raft log flush optimization
Before this change, leader and follower log flushes were sequential.
With this change, the leader flushes after it sends append requests to
followers and it might even commit if the majority followers reply
before it flushes.

This optimization is discussed in the Raft dissertation Section 10.2.1.

This commit also includes other minor fixes and improvements.