Implement WAN protocol
Added support for setting, transmitting and reading the negotiated WAN
protocol version during the serialisation/deserialization process. We
reuse the existing version field and bytes used for intra-cluster
versioning (RU) but use negative values to denote WAN protocol version.
As such, it is not possible to set both the WAN protocol version and
cluster version for RU on a single versioned instance, which is not a
real restriction as the instance will either be sent to a cluster member
or a WAN target, not both at the same time.

Also did some minor refactoring such as moved some methods and extracted
inner classes.