Improve data consistency for IMap and ICache iterators
Improves consistency of returned data for IMap and ICache iterators
in the face of concurrent mutation of map and cache data and migration
of data between members.

The new contract for the iterator:
- keys which have been deleted or added since the iteration started
must be observed at most once, other keys must be observed exactly once
- the next method may return an entry with a null value. This can happen
when prefetchValues is false and the key was present when the entry
batch was fetched during the consumption of this batch, the entry was
removed.

Fixes: https://github.com/hazelcast/hazelcast/issues/11106