Publish metrics to MC and JMX
Metrics are recorded to a ringbuffer that Management Center can read out
by using ReadMetricsMessageTask. The task reads the metrics recorded
since the last seen sequence number. The next sequence number that MC
can continue reading the Metrics with is returned in the response. The
ringbuffer, the ReadMetricsOperation and the task are almost identical
copies of the similar Jet classes.

Besides the described mechanism used for recording and publishing,
configuring the metrics system has also changed. Before this commit,
metrics could be configured with JVM arguments, now it can be done with
MetricsConfig both on the member and on the client side. This
configuration is shared with the diagnostic's MetricsPlugin.

The client protocol changes required for the ReadMetricsMessageTask is
also include in this commit.