Demote Raft leader if the majority remains silent
If the Raft leader node receives no append entries response from
the majority within the leader heartbeat timeout, it demotes itself to
the follower role and notifies the futures of the pending operations.
This is done to not to hang the callers for a long duration if
the leader remains partitioned away from the majority.