Rename cluster name in AbstractWanEventBase
We send WAN target information to MC as `targetClusterName` in
WAN events. The name may be misleading, since we put the publisherId
to this field that can either be the group name or a publisher id
set in the config. We change the name of the field to `wanPublisherId`
to be consistent.

This is noticed since recent refactoring changed targetGroupName to
targetClusterName that led to rendering `undefined` targets in MC.