Fix for OperationTimeoutException when join request is sent to non-master (#1661)
When a join request is sent to non-master, it is automatically routed to the
master node. However if the master node is to leave the cluster then the join
request can be failed early with a OperationTimeoutException. The fix is
to make a join request never timeout - as the job can be potentially running for
a long time.

Fixes #1660