Fixes back pressure enabled due to write through
The problem was that write through requires the
CallIdSequenceBackpressureCounter because it needs
to know the number of inflight invocations to
determine if the system is concurrent or not.

However it will then pick up the default number of
invocations that should only be picked up when
back pressure is enabled.

so this fixes that; if back pressure is disabled, then
Integer.MAX_VALUE is the maximum inflight invocations
whic is the same as saying there is no back pressure.