Changes test Invocation_ExceptionTest to avoid usage of OOME
OutOfMemoryException triggers the default OOME handler that
shuts down all HazelcastInstances known to the current JVM, so a
side-effect can be that the InvocationFuture of an operation
that throws OOME is completed with HazelcastInstanceNotActiveException.
In order to test wrapping of Errors in InvocationFuture, instead of
OOME we now use ExceptionInInitializerError.

Also fixes some copy-paste errors in comments