Add MC Operations to IMDG Client (#15662)
MC will be using IMDG for reading metrics as well as sending operations to cluster members. This PR is for adding client support for the following three operations that MC can send to a cluster:

* change cluster state
* get map config
* update map config

This is the first step in migrating all operations. Once this gets merged, a separate PR will follow that migrates the remaining operations. The cleanup of old tasks will also follow in a separate PR.

This PR also:
* moves `readMetrics` operation under the client side `ManagementCenterService`
* fills the `ReadMetricsMessageTask#getMethodName` implementation

Client protocol PR: https://github.com/hazelcast/hazelcast-client-protocol/pull/237