Rename CPSubsystemManagementService.restart() to .reset()
We have CPSubsystemManagementService.restart() to wipe-out and reset
the whole CP Subsystem in case of permanent loss of majority. After CP
Subsystem Persistence work is done, it turned out that .restart() is a
bad choice for this functionality because it creates confusion with
restarting CP members by using persistence. This PR renames this method
to .reset() to eliminate this confusion.