Fix possible metrics (de)registration livelock (#15777)
* Fix possible metrics (de)registration livelock

The `incrementMod` method used a get+CAS in an infinite loop. This is
prone to a livelock when many threads contend to run it. The commit
updates this with incrementing an AtomicInteger which uses the CPU's
fetch-and-add instruction and should not be susceptible to this.

Fixes #15503

* Fix invalid exit code

Negative numbers aren't allowed, -1 was interpreted as 255.