Refine member metrics configuration
This change refines the member-side metrics configuration to address two
issues with the previously added config:
- It was not clear that `retention-seconds` was for MC only
- Further improving the MC and JMX configs was not supported by the
attributes

To address these, the configuration changes as follows.

Before this change:
```
<metrics enabled="false" mc-enabled="false" jmx-enabled="false">
    <collection-interval-seconds>10</collection-interval-seconds>
    <retention-seconds>30</retention-seconds>
    <metrics-for-data-structures>true</metrics-for-data-structures>
    <minimum-level>DEBUG</minimum-level>
</metrics>
```

After this change:
```
<metrics enabled="false">
    <management-center enabled="false">
        <retention-seconds>30</retention-seconds>
    </management-center>
    <jmx enabled="false"/>
    <collection-frequency-seconds>10</collection-frequency-seconds>
    <data-structure-metrics-enabled>true</data-structure-metrics-enabled>
    <level>DEBUG</level>
</metrics>
```

The programmatic and YAML configuration follows the same structure.

It is also a requirement that all metrics configurations should
overrideable via system properties. This is also implemented in this
change by overriding the configuration from the properties during the
`Node` instances is created. The concrete system properties are added to
the javadocs of the metrics config classes and to the xsd too.

Besides refining the changes above, the following gaps were also closed:
- Extends javadoc and xsd documentation for the metrics config elements
- Add Spring XML configuration
- Add full-example config comments
- Add metrics config to hazelcast-fullconfig.[xml|yaml]
- Add metrics metrics config to hazelcast-default.[xml|yaml]