Fail with NotLeaderException when RaftNode is terminated.
When RaftNode is terminated, from its perspective it is not always
certain that if its Raft group was terminated or its container (JVM or
Hazelcast instance) is shutting down. Because of this, RaftNode responds
to replicate and query requests with NotLeaderException, simply meaning
that it cannot handle that request. The Raft invocation retry mechanism
will check if the Raft group is terminated therefore the request must be
failed, or the Raft group is still active and the request must be
forwarded to another Raft node.