Apply Raft node management fixes
* As a liveliness optimization, we are triggering Raft node creations
inside MetadataRaftGroupManager when a new Raft group is created.
However, once we have persistence and restarts, this is not a safe
optimization anymore because if we are replaying Raft log entries after
a restart, we might be re-creating Raft nodes for destroyed Raft groups.
These orphan Raft nodes might elect their leader and handle incoming
requests -BOOM!- until their destroy operation is replayed
in the Metadata CP Group. This optimization is now moved to a safe
place: RaftInvocationManager.
* When a Raft Node is being terminated, its CP group id must be added to
the "destroyed group ids set" only if it is certainly known that this
Raft node's CP group is terminated. For instance, when a CP member is
shutting down, it also terminates its Raft nodes and it should not fail
incoming requests with CPGroupDestroyedException during its shutdown.
* Destroy Raft node state store on Raft node creation if it is learnt
that the Raft group is already destroyed. This logic was mistakenly
removed in 1937db9ba17a2d2b54be2425048149ed9befd56f and now it is added
back with this commit.