Fix WAN queue migration for structures with backupCount>1
Currently, on migration commit/rollback, we were clearing all partition
queues if the new replica index was greater than 1. This is fine if the
backup count for a map/cache is 1 but if the backup count is higher, we
clear the queues even though we shouldn't.
We now clear queues depending on the new replica index and the queue
backup count, exactly how map service does it.
As a side-effect, we simplify the WAN publisher SPI since we can now
remove two methods from the SPI - one which was even a leaking
abstraction mentioning "queues".

Also, fixed NPE issues in CacheMergeOperation when backupCount is 0 and
renamed "WanEventQueueMigrationListener" which is part of the publisher
SPI to remove mentioning of "queues".

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2091