Handle destroyed Raft group on Raft node init task
When a queried Raft node is not existing on local member, the local
member asks the METADATA CP group to get the CP group info and create
the Raft node if needed. If the CP group is already destroyed, the local
member must process this information using the Raft node termination
logic by calling `RaftService.terminateRaftNode()`.