EntryProcessorOffloadableTest: made tests less susceptible to hiccups
The test testHeartBeatsComingWhenEntryProcessorOffloaded
uses a lower value (8s) for OPERATION_CALL_TIMEOUT_MILLIS
so that operation heartbeats are sent more frequently.

However, OPERATION_CALL_TIMEOUT_MILLIS was lowered in
a @Before method. As a result, it applied to all tests,
making them susceptible to timeouts due to hiccups etc.

Made sure that the reduced OPERATION_CALL_TIMEOUT_MILLIS
is used only in testHeartBeatsComingWhenEntryProcessorOffloaded.

https://github.com/hazelcast/hazelcast-enterprise/issues/3303