Catch and log exceptions from dynamic metrics providers
If the dynamic metrics collection encounters an exception it bubbles the
exception up to the caller triggered the periodic task that triggered
the collection. It is fixed by catching and logging the exception
without bubbling it up.

Throwing an NPE from
`TcpIpNetworkingService.MetricsProvider#provideDynamicMetrics` in
advanced networking mode is fixed by adding a null check around
`unifiedEndpointManager`.

Fixes #15932