Send client metrics to the member and merge them
The metrics collected on the client side are sent to the member to
include them in the metrics blob collected on the member side. The
metrics are sent from the client side to the member side using the same
metrics compression that the member side is using.

If enabled in the metrics configuration, the client exposes the collected
client-side metrics automatically on JMX under
`com.hazelcast/Metrics/$clientName`.

Adds two client-side dynamic metric providers:
- NearCacheMetricsProvider
- ClusterConnectionMetricsProvider