Remove configuration for user-defined services SPI
In 4.0, we removed support for adding custom services by moving the
NodeEngine and all services to internal packages. With this, we remove
the configuration part for user-defined services. We still allow adding
custom services through a private ConfigAccessor method which is useful
for some tests or for users wanting to use private API at their own risk.

Also did some JDK8 cleanup.