MC: Add getTimedMemberState operation to client (#15949)
* MC: Add getTimedMemberState operation to client

MC 4.0 will not use the new metrics due to a recent change in MC 4.0
scope. For this reason, we need to get the TimedMemberState object from
the members using the client.

* MC 4.0 will use this operation to get the timed member state from
IMDG 4.0.
* MC 4.1 will use `readMetricsAsync` to get the new metrics from
IMDG 4.0 and 4.1.

Hence, this operation will only be functional in 4.0.x and removed in
4.1.

Note 1: We keep `master` branch sending `TimedMemberState` to MC as
we're developing the cluster communication model change functionality
in a separate feature branch and we need to keep the `master` branch of
IMDG and MC compatible for now. Removing the code that sends
`TimedMemberState` to MC will happen in a separate PR when all MC
operations are performed using the client on `master` branch.

Note 2: Members still need to send their `TimedMemberState` to `MC` for
client filtering to work. Client filtering will be update to not require
 it in a separate PR.