Prevent concurrent migration finalization to cancel scheduled tasks
A FinalizeMigrationOperation can be invoked concurrently to other
partition migrations. As such, the operation may invoke beforeMigration
and commitMigration/rollbackMigration methods while there is a
concurrent invocation to prepareReplicationOperation for another
partition.
This may cause the scheduled executor service to mistakenly think it is
preparing a replication operation for a migration while in fact there
was is no migration.

Also, did some JDK8 cleanups and made all quorum tests to use small
test instances.

Fixes:
https://github.com/hazelcast/hazelcast/issues/15904
https://github.com/hazelcast/hazelcast/issues/12197