Revising connection retry config
ConnectionRetryConfig was not allowing retry with fixed amount of time
and shutdown after some time.

Deprecated/Removed connectionAttemptLimit and period was solution
to this.

In 4.0 since we want single place to configure this,
ConnectionRetryConfig is modified to support both exponential and
fixed wait time.

To achieve this:
fail-on-maxbackooff is removed
cluster-connect-timeout-seconds is added.(default is 20)